Rename `config` to `tsconfig`
